backend/requirements.txt
django
djangorestframework
django-cors-headers
ccxt
yfinance
pandas-ta
pandas
python-dotenv
scikit-learn
numpy
